+++ Assets/Project/TitleScenes/Scripts/GameSystem.cs
@@ -5,57 +5,61 @@ using System.Linq;
 using System.Collections;
 using UnityEngine.SceneManagement;
 using UnityEngine.XR;
+using VrScene;
 
-public class GameSystem : MonoBehaviour
+namespace TitleScene
 {
-    public GameObject ModeSelectPanel;
-    public GameManager GameManager;
-    public ToggleGroup toggleGroup;
+    public class GameSystem : MonoBehaviour
+    {
+        public GameObject ModeSelectPanel;
+        public GameManager GameManager;
+        public ToggleGroup toggleGroup;
 
-    private void Awake()
-    {
-        XRSettings.enabled = false;
-        ModeSelectPanel.SetActive(false);
-    }
-    public void SelectGameMode()
-    {
-        ModeSelectPanel.SetActive(true);
-    }
-
-    private void Update()
-    {
-        if (Application.platform == RuntimePlatform.Android)
+        private void Awake()
         {
-            if (Input.GetKeyDown(KeyCode.Escape))
+            XRSettings.enabled = false;
+            ModeSelectPanel.SetActive(false);
+        }
+        public void SelectGameMode()
+        {
+            ModeSelectPanel.SetActive(true);
+        }
+
+        private void Update()
+        {
+            if (Application.platform == RuntimePlatform.Android)
             {
-                ModeSelectPanel.SetActive(false);
+                if (Input.GetKeyDown(KeyCode.Escape))
+                {
+                    ModeSelectPanel.SetActive(false);
+                }
             }
         }
-    }
 
-    public void OnClickWorldSelectButton(string ID)
-    {
-        // BlockManagerにIDを渡す
-        BlockManager.WorldID = ID;
+        public void OnClickWorldSelectButton(string ID)
+        {
+            // BlockManagerにIDを渡す
+            BlockManager.WorldID = ID;
 
-        // GameManagerにModeを渡す
-        string selectedLabel = toggleGroup.ActiveToggles()
-            .First().GetComponentsInChildren<Text>()
-            .First(t => t.name == "Label").text;
+            // GameManagerにModeを渡す
+            string selectedLabel = toggleGroup.ActiveToggles()
+                .First().GetComponentsInChildren<Text>()
+                .First(t => t.name == "Label").text;
 
-        GameManager.Mode = selectedLabel.Replace("Mode", "");
+            GameManager.Mode = selectedLabel.Replace("Mode", "");
 
-        // VrSceneを読み込む
-        SceneManager.LoadScene("Vr");
-        Screen.orientation = ScreenOrientation.LandscapeLeft;
+            // VrSceneを読み込む
+            SceneManager.LoadScene("Vr");
+            Screen.orientation = ScreenOrientation.LandscapeLeft;
 
-    }
+        }
 
-    public void MoveSetting()
-    {
-        SceneManager.LoadScene("SettingScene");
+        public void MoveSetting()
+        {
+            SceneManager.LoadScene("SettingScene");
 
-        Screen.orientation = ScreenOrientation.LandscapeLeft;
+            Screen.orientation = ScreenOrientation.LandscapeLeft;
 
+        }
     }
 }
